+++ serve/src/gui.html
@@ -48,24 +48,9 @@
                                 </el-option>
                             </el-select>
                             <span><a :href="data.configLinks[state.pretrainedModel]" target="_blank">Learn more...</a></span>
-                            <!--
-                            <sly-field title="Code"
-                                   style="padding-top: 0; padding-bottom: 0; margin: 10px 5px;">
-
-                                <sly-icon slot="icon" :options="data.github_icon"></sly-icon>
-                                <span><a :href="data.pretrainedModels[state.pretrainedModel]['code_url']" target="_blank">{{state.pretrainedModel}}</a></span>
-                            </sly-field>
-                            <sly-field title="Paper"
-                                   style="padding-top: 0; padding-bottom: 0; margin: 10px 5px;">
-
-                                <sly-icon slot="icon" :options="data.arxiv_icon"></sly-icon>
-                                <span><a :href="data.pretrainedModels[state.pretrainedModel]['paper_url']" target="_blank">{{data.pretrainedModels[state.pretrainedModel]['paper_title']}}</a></span>
-                            </sly-field>
-                            -->
                         </div>
                         <div>
                             <div class="mb10">Pretrained model:</div>
-                            <!-- <div style="font-size: 11px;color: #848484;" class="mb10">See more details in <a :href="data.pretrainedModels[state.pretrainedModel]['config_url']" target="_blank">model config files</a>. </div> -->
                             <table class="beautiful-table">
                                 <thead>
                                 <tr>
