+++ train/src/ui/hyperparameters.html
@@ -94,7 +94,7 @@
             <sly-field title="Optimizer">
                 <div slot="description">
                     Choose optimizer and its settings, learn more in official
-                    <a href="https://pytorch.org/docs/1.9.0/optim.html#algorithms" target="_blank">
+                    <a href="https://pytorch.org/docs/1.10.0/optim.html#algorithms" target="_blank">
                         pytorch documentation
                     </a>
                 </div>
@@ -165,7 +165,7 @@
                 <el-switch v-model="state.useClassWeights" on-color="#13ce66" off-color="#B8B8B8" :disabled="data.done6"></el-switch>
             </sly-field>
             <sly-field v-if="state.useClassWeights" title="Manual class weights" description="Provide class weights for each class (floats splitted by comma)">
-                <div>Selected classes order: ( <span v-if="state.selectedClasses.includes(class_name)" v-for="class_name in data.classesList" style="font-style: italic;">{{class_name}} </span><span style="font-style: italic;">__bg__</span> )</div>
+                <div>Selected classes order: ( <span v-if="state.selectedClasses.includes(class_name)" v-for="class_name in data.classesList" style="font-style: italic;">{{class_name}},</span><span style="font-style: italic;">__bg__</span> )</div>
                 <el-input v-model="state.classWeights" placeholder="1.2,0.4,0.3"  :debounce="1000" :disabled="data.done6"></el-input>
             </sly-field>
 
@@ -214,7 +214,7 @@
         <el-tab-pane label="Learning rate scheduler (Advanced)" name="lr_scheduler" :disabled="data.done6">
             <div class="mb10">
                 Choose LR scheduler and its settings, learn more in official
-                <a href="https://pytorch.org/docs/1.9.0/optim.html#how-to-adjust-learning-rate" target="_blank">
+                <a href="https://pytorch.org/docs/1.10.0/optim.html#how-to-adjust-learning-rate" target="_blank">
                     pytorch documentation
                 </a>
                 or in
