+++ train/src/ui/monitoring.html
@@ -63,7 +63,6 @@
             <sly-icon slot="icon" :options="{ color: '#006ad4', bgColor: '#bbe2ff', rounded: false }">
                 <i class="zmdi zmdi-trending-up"></i>
             </sly-icon>
-<!--            data.chartLR.options.smoothingWeight = val;-->
             <el-slider
                     :value="state.smoothing"
                     :min="0" :max="0.95" :step="0.05" show-input
@@ -77,7 +76,6 @@
         </sly-field>
 
         <sly-field title="Main training charts" description="Real-time training metrics">
-            <div>{{state.chartTrainLoss.options.smoothingWeight}}</div>
             <div class="fflex">
                 <sly-line-chart
                         style="width: 35%;"
